requirements.txt
wheel
setuptools
twine
flake8
requests
pyyaml
